third_party/WebKit/LayoutTests/inspector/tracing/timeline-layout-with-invalidations-expected.txt

Tests the Timeline API instrumentation of layout events with invalidations.

PASS
 

Running: testLocalFrame
first layout invalidations[
    {
        cause : {reason: Style changed, stackTrace: .../inspector/tracing/timeline-layout-with-invalidations.html:11}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElement'"
        selectorPart : undefined
        type : "LayoutInvalidationTracking"
    }
]
second layout invalidations[
    {
        cause : {reason: Style changed, stackTrace: .../inspector/tracing/timeline-layout-with-invalidations.html:13}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElement'"
        selectorPart : undefined
        type : "LayoutInvalidationTracking"
    }
]

Running: testSubframe
first layout invalidations[
    {
        cause : {reason: Style changed, stackTrace: .../inspector/tracing/timeline-layout-with-invalidations.html:23}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV"
        selectorPart : undefined
        type : "LayoutInvalidationTracking"
    }
]
second layout invalidations[
    {
        cause : {reason: Style changed, stackTrace: .../inspector/tracing/timeline-layout-with-invalidations.html:25}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV"
        selectorPart : undefined
        type : "LayoutInvalidationTracking"
    }
]
